chore: creating routes' folder and file
