+++ package.json
@@ -34,6 +34,7 @@
   },
   "dependencies": {
     "@fastify/cors": "^8.2.0",
+    "@fastify/jwt": "^6.5.0",
     "@prisma/client": "^4.10.1",
     "fastify": "^4.12.0",
     "zod": "^3.20.6"
