+++ src/lib/prisma.ts
@@ -1,0 +1,2 @@
+import { PrismaClient } from '@prisma/client';
+export default new PrismaClient();
