+++ src/model/userModel.ts
@@ -1,5 +1,4 @@
-import { PrismaClient } from '@prisma/client';
-const prisma = new PrismaClient();
+import prisma from '../lib/prisma';
 
 export default class UserModel {
   async selectClientById(userID: number) {
